20210221-222517

+++ scripts/reshape_rep.py
@@ -17,20 +17,37 @@ def reshape(args, params, filenames):
     log.logger.debug('started')
     try:
         fa=parse_fasta(args.rep)
+        if len(fa) == 0:
+            log.logger.error('No sequence found in your input rep file (-rep flag).')
+            exit(1)
         fa_keep={}
         fa_unknown={}
         known_name_to_clas={}
+        rep_kept_n=0
+        rep_removed_n=0
         for header in fa:
             hs=header.split('\t')
             if len(hs) == 3:
                 if not hs[1] in args.rep_headers_to_be_removed:
                     fa_keep[header]=fa[header]
                     known_name_to_clas[hs[0].replace('>', '')]=hs[1]
+                    rep_kept_n += 1
+                else:
+                    rep_removed_n += 1
             elif len(hs) == 1:
                 fa_unknown[header]=fa[header]
+                rep_kept_n += 1
+            else:
+                log.logger.error('Fasta header of your input rep file (-rep flag) does not look like the appropriate fasta file conveted from Dfam.')
+                exit(1)
+        if rep_kept_n == 0:
+            log.logger.error('All sequences in your input rep file (-rep flag) match with non-MEs listed in a file specified with -repremove flag.')
+            exit(1)
+        else:
+            log.logger.info('N=%d repeats found in %s. N=%d will be analyzed. N=%d will be excluded due to non-ME repeats.' % (len(fa), args.rep, rep_kept_n, rep_removed_n))
         with open(filenames.reshaped_rep, 'w') as outfile:
             for header in fa_keep:
-                outfile.write(header +'\n'+ fa_keep[header] +'\n')
+                outfile.write('%s\n%s\n' % (header, fa_keep[header]))
             outfile.flush()
             os.fdatasync(outfile.fileno())
         dels=set()
@@ -41,31 +58,35 @@ def reshape(args, params, filenames):
         if len(dels) >= 1:
             for d in dels:
                 del(fa_unknown[d])
-        with open(filenames.rep_unknown_fa, 'w') as outfile:
-            for header in fa_unknown:
-                outfile.write(header +'\n'+ fa_unknown[header] +'\n')
-            outfile.flush()
-            os.fdatasync(outfile.fileno())
-        blastn.makeblastdb(filenames.reshaped_rep, filenames.repdb)
-        blastn.blastn_for_unknown_rep_ident(args, params, filenames.rep_unknown_fa, filenames.repdb, filenames.blast_tmp_res)  # determine TE class of unknown rep
-        hits={}
-        with open(filenames.blast_tmp_res) as infile:
-            for line in infile:
-                ls=line.split()
-                if not ls[0] == ls[1]:
-                    if not ls[0] in hits:
-                        hits[ls[0]]=ls[1]
-        for header in hits:
-            hit=hits[header]
-            if hit in known_name_to_clas:
-                fa_keep['>%s\t%s' % (header, known_name_to_clas[hit])]=fa_unknown['>%s' % header]
-        with open(filenames.reshaped_rep, 'w') as outfile:
-            for header in fa_keep:
-                outfile.write(header +'\n'+ fa_keep[header] +'\n')
-            outfile.flush()
-            os.fdatasync(outfile.fileno())
-        os.remove(filenames.rep_unknown_fa)
-        os.remove(filenames.blast_tmp_res)
+        if len(fa_unknown) >= 1:
+            log.logger.debug('%d MEs with ambiguous subclass found.' % len(fa_unknown))
+            with open(filenames.rep_unknown_fa, 'w') as outfile:
+                for header in fa_unknown:
+                    outfile.write('%s\n%s\n' % (header, fa_unknown[header]))
+                outfile.flush()
+                os.fdatasync(outfile.fileno())
+            blastn.makeblastdb(filenames.reshaped_rep, filenames.repdb)
+            blastn.blastn_for_unknown_rep_ident(args, params, filenames.rep_unknown_fa, filenames.repdb, filenames.blast_tmp_res)  # determine TE class of unknown rep
+            hits={}
+            with open(filenames.blast_tmp_res) as infile:
+                for line in infile:
+                    ls=line.split()
+                    if not ls[0] == ls[1]:
+                        if not ls[0] in hits:
+                            hits[ls[0]]=ls[1]
+            for header in hits:
+                hit=hits[header]
+                if hit in known_name_to_clas:
+                    fa_keep['>%s\t%s' % (header, known_name_to_clas[hit])]=fa_unknown['>%s' % header]
+            with open(filenames.reshaped_rep, 'w') as outfile:
+                for header in fa_keep:
+                    outfile.write(header +'\n'+ fa_keep[header] +'\n')
+                outfile.flush()
+                os.fdatasync(outfile.fileno())
+            os.remove(filenames.rep_unknown_fa)
+            os.remove(filenames.blast_tmp_res)
+        else:
+            log.logger.debug('No MEs with ambiguous subclass found.')
     except:
         log.logger.error('\n'+ traceback.format_exc())
         exit(1)
